refine IT tests
