+++ ingest-src/src/test/resources/resources/ingestConfig.yaml
@@ -19,12 +19,12 @@ ingest-info:
     ingestLock: "/mrt.lock"
 
   store-info:
-    store.1: "http://store:8080/store"
-    access.1: "http://store:8080/store"
-    localID: "http://it-server:{!ENV: it-server.port !DEFAULT: 4567}"
+    store.1: "http://localhost:4567"
+    access.1: "http://localhost:4567"
+    localID: "http://localhost:4567"
 
   queue-info:
-    QueueService: "zoo:2181"
+    QueueService: "localhost:{!ENV: it-zoo.port !DEFAULT: 2181}"
     QueueName: "/ingest"
     IngestQNames: "ingest"
     InventoryName: "/mrt.inventory.full"
